// backend/src/common/enums/travel-preference.enum.ts
export enum TravelPreference {
  REMOTE_ONLY = 'REMOTE_ONLY',
  WILLING_TO_TRAVEL_HOME_LOCATION = 'WILLING_TO_TRAVEL_HOME_LOCATION',
  WILLING_TO_TRAVEL_FIRE_ZONE = 'WILLING_TO_TRAVEL_FIRE_ZONE',
  WILLING_TO_TRAVEL_FIRE_CENTRE = 'WILLING_TO_TRAVEL_FIRE_CENTRE',
  WILLING_TO_TRAVEL_REGION = 'WILLING_TO_TRAVEL_REGION',
  WILLING_TO_TRAVEL_ANYWHERE = 'WILLING_TO_TRAVEL_ANYWHERE',
  WILLING_TO_TRAVEL_UNKNOWN = 'WILLING_TO_TRAVEL_UNKNOWN',
}

export enum BcwsTravelPreference {
  REMOTE_ONLY = 'REMOTE_ONLY',
  WILLING_TO_TRAVEL_HOME_LOCATION = 'WILLING_TO_TRAVEL_HOME_LOCATION',
  WILLING_TO_TRAVEL_FIRE_ZONE = 'WILLING_TO_TRAVEL_FIRE_ZONE',
  WILLING_TO_TRAVEL_FIRE_CENTRE = 'WILLING_TO_TRAVEL_FIRE_CENTRE',
  WILLING_TO_TRAVEL_ANYWHERE = 'WILLING_TO_TRAVEL_ANYWHERE',
  WILLING_TO_TRAVEL_UNKNOWN = 'WILLING_TO_TRAVEL_UNKNOWN',
}

export enum EmcrTravelPreference {
  REMOTE_ONLY = 'REMOTE_ONLY',
  WILLING_TO_TRAVEL_HOME_LOCATION = 'WILLING_TO_TRAVEL_HOME_LOCATION',
  WILLING_TO_TRAVEL_REGION = 'WILLING_TO_TRAVEL_REGION',
  WILLING_TO_TRAVEL_ANYWHERE = 'WILLING_TO_TRAVEL_ANYWHERE',
  WILLING_TO_TRAVEL_UNKNOWN = 'WILLING_TO_TRAVEL_UNKNOWN', 
}